backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
easyAI==1.0.0.4
numpy==2.4.1
pillow==12.1.0
pygame==2.6.1
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
psycopg2-binary>=2.9